Clean out old cruft and move to Cask
After years of people editing this repo only for themselves, it has largely been abandoned because it no longer works for anyone. I still think it's a good idea, so I'm trying to revitalize it even if I'm the only person that's going to use it.

My first step is to clean out all of the old cruft. A lot of it doesn't work well or interoperate anymore and debugging and Emacs config this large is pretty tough. There are a lot of good tools available in 2016 that weren't available when we started this project in 2002, so I'm removing a bunch of things and backporting anything I find that seems like it's still a good idea.

The first idea is Cask. Cask is a Bundler-alike for emacs. In the olden days, there was nothing like Bundler so our best move was to commit every package we wanted to have. We knew that was bad and it sucked in a lot of ways. It also contributed to this repo becoming a mess.

If you want to get started with this repo, run `cask install` from the .emacs.d directory.