Stop using Apple's deprecated java home setting mechanism
Why is this change needed?
--------------------------
Once upon a time, Apple had a dream of it being easy to set and manage your various javas. One part of it was /usr/libexec/java_home which would determine where your java home directory was. The Java community really doesn't like this and has been working against it for years. It's no longer safe to rely on. Best case, it prints an error message about not having any java versions installed because we're using asdf to manage our java stuff now. Setting of JAVA_HOME is now done in bash_profile where we confirm the existence of asdf on your machine.

How does it address the issue?
------------------------------
I deleted it.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
You can see the many links in script/setup for deeper knowledge of the ongoing fight among the java community about environment management. It's kind of a crappy read.