Add the Java bin directory to your path when JAVA_HOME is set.
Why is this change needed?
--------------------------
This was previously taken care of by Apple's Java ecosystem thing which can no longer be relied on. It should be taken care of by asdf's java plugin, but that also doesn't do the right thing. It would be good to fix the asdf java plugin, but I don't have time for that right now.

In any case, Instrumental needs the java bin directory in the path to run sbt things during project setup. It probably also needs it at other times.

How does it address the issue?
------------------------------

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/skotchpine/asdf-java/issues/46