Install dropbox and prompt you to start syncing your code and other important things
Why is this change needed?
--------------------------
If you're keeping important things in Dropbox, like your repos or secrets, you'll want to start syncing early so they're done close to the end of the script running.

How does it address the issue?
------------------------------

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------