Switch to using the EB standard setup script location and file name
Why is this change needed?
--------------------------
Behaviornauts expect, by convention, to find the setup script for our projects in the `script` directory in the root of the project and for it to be named `setup` with no file extension.

Also, the script/setup file has eclipsed the intended purpose of the old `setup.sh` file.

How does it address the issue?
------------------------------
Executes the new file from the expected place.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------