Add some comments about further work
Why is this change needed?
--------------------------
There's a lot more that should happen in the setup script, but I didn't have time to do it all. These are some notes about we may want to add. In general, there's a lot of software that gets installed that needs to be set up. We should prompt that setup while the process is fresh and get everything out of the way.

How does it address the issue?
------------------------------

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------