Give emacs a monospace font by default
Why is this change needed?
--------------------------
It's crazy to code in non-monospace fonts

How does it address the issue?
------------------------------
Sets the default to a font that everybody should have. I don't love it, but it's better than no config. I should really get a better font than this, though.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------