Add some conveniences for switching windows
Why is this change needed?
--------------------------
Switching windows is too common an action to require C-x o.

How does it address the issue?
------------------------------
M-p for next window. M-o for previous window. C-x O for switch-window which will let you switch to directly to an arbitrary window.

I realize it seemingly makes more sense for 'previous' to be bound to 'p', but when you're actually using the keys 'p' is to the right and that's the direction the cursor moves when that key is pressed. I deemed it more intuitive for the relation of the keys to match the relation of the window movement.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------