Remove the elpa archives from the repo
Why is this change needed?
--------------------------
The prior config depended on having the emacs packages in the repo. The new use-package based configuration automatically downloads and installs what you need. There's no need to have a bunch of (probably outdated) packages in the repo. Also, having an outdated archive can interfere with loading dependencies. The error messages for that state are obtuse and made me sad.

How does it address the issue?
------------------------------
Deletes the unused archive.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------