Add basic config for project management and rapid emacs navigation
Why is this change needed?
--------------------------
These are common actions and should be as fast as possible.

How does it address the issue?
------------------------------
Installs Projectile for project management and Helm for emacs navigation. Projectile seems to be the default choice, so that's probably not surprising. Helm and Ivy are fairly competitive, so I'll talk about that. I tried both and liked some things about both. I decided on Helm because the default choices when using projectile-find-file were better. I did common searches in Instrumental and DocRaptor. Ivy would often return a large list of useless files the first time until I typed almost the entire file name. As an example, searching for the yml file that defines part of the CloudWatch integration. Ivy would return a lot of icon files we use to display which various AWS services we support before the yml file. Helm showed the yml file in the first five results. That was typical of my testing, so I went with Helm.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://www.sandeepnambiar.com/my-minimal-emacs-setup/