Refactor secret env variables
Turns out the gpg agent was unnecessary, and was causing issues with
secrets being loaded. I think the old code was reloading keys, so while
they weren't loaded the secrets were failing to load.

Pinentry and gpg work together with Keychain Access to allow decryption,
since the passphrase is saved in Keychain Access.