Create a shortcut for switching projects
Why is this change needed?
--------------------------
It can be a little tedious to navigate the filesystem to a different projet root.

How does it address the issue?
------------------------------
Projectile find-files remembers previous projects, so you should be able to just select from those once you've gone to a project before.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------