Print a warning message if open file limits are less than 200k
Why is this change needed?
--------------------------
Some of the test files in Instrumental require more open files than that to run succesfully and the error you get if you don't have it is extremely obtuse (no error message, just stops running).

How does it address the issue?
------------------------------
Print a warning when you open a terminal and gives directions on how to fix the problem.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

Co-authored-by: Alex Overbeck <alex@expectedbehavior.com>