Move keybindings for rapid buffer switching onto the use-package definition
Why is this change needed?
--------------------------
If you put keybindings outside the use-package and the package fails to load for some reason, then you get to deal with errors about trying to bind keys to functions that don't exist.

How does it address the issue?
------------------------------
use-package will only try to bind the keys if the package actually loads, so no extra binding errors if it doesn't

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------