Moves frame switching into use-package and removes the last of the global keybindings for window movement
Why is this change needed?
--------------------------
Same as last time. Global keybindings outside use-package can cause errors. Also, makes it harder to keep track of the purpose and relationship of the bindings.

How does it address the issue?
------------------------------
Moves the bindings to the package.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------