Stop displaying the default scratch buffer message and change the scratch major mode to Ruby
Why is this change needed?
--------------------------
It's just a slight nicety. We don't need a message telling us what a scratch buffer is. Ruby is the most common language at Expected Behavior right now, so it's a convenient mode to start with.

How does it address the issue?
------------------------------
You basically get a blank Ruby file out of the box. If you just want to fire up emacs and start pounding out Ruby, it's now as easy as it can get.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------