Enable looking up documentation in Dash using C-c d
Why is this change needed?
--------------------------
I've been using Dash as my primary way of looking up documentation for a while now and find it pretty helpful. This is just a convenience to get there a little faster.

How does it address the issue?
------------------------------
Installs the dash-at-point package from MELPA and binds the suggested key.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/stanaka/dash-at-point