Add keystrokes for quickly navigating to Instrumental and DocRaptor
Why is this change needed?
--------------------------
I go to those projects all the time, want it to be fast, and don't want to leave emacs to do it.

How does it address the issue?
------------------------------
Add a key combo along the lines of projectiles file navigation keys that opens the README for the given project. Should be easy to get where you want to go from there.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------