Allow navigation between buffers using the same keystroke most OSX applications respond to
Why is this change needed?
--------------------------
I've used OSX for so many years that command-shift- { or } is ingrained in me for switching rapidly between the "tabs" of a thing. Since it's something I do lot, continuing to use the full buffer switcher seems like a needless waste of time.

How does it address the issue?
------------------------------
The standard OSX keystrokes are remapped to the next and previous buffer commands.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------