Force TLS 1.3
Why is this change needed?
--------------------------
GnuTLS currently has issues that will cause it to incorrectly negotiate the TLS version, causing package installations to fail with an obtuse "not found" message.

How does it address the issue?
------------------------------
Tell Emacs to use TLS 1.3 by default.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
Added in a comment above the code.