Put transient files in their own directory
Why is this change needed?
--------------------------
By default, Emacs puts these files in the current working directory. It's a bit irritating when looking at the Git situation.

How does it address the issue?
------------------------------
Sets a bunch of common cruft files to be put into a cache in the home directory. I probably didn't get them all, but it's a start.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------