Only try to initialize packages once
Why is this change needed?
--------------------------
It's sometimes convenient to re-eval the entire emacs config and this thing complains if it's already been initialized.

How does it address the issue?
------------------------------
Check to see if it's already been initialized. If not, do it.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------