Add some common interface conveniences
Why is this change needed?
--------------------------
These are things I do all the time anyway, so why not automate them?

How does it address the issue?
------------------------------
* remove as much stuff from the UI as possible
It's just a bunch of visually distracting chrome. Kill it.

* Switch from global-linum-mode to global-display-line-numbers-mode
For both performance and visual reasons. GLM didn't do well on long files and caused visible slowdowns. Also, it shrank the font size as the length of the line numbers increased. It was very hard to read on files with over a 999 lines.

Line numbers are disabled for modes where it doesn't make sense.

* Fullscreen on startup
I always use Emacs fullscreen. No need for the extra keystroke.

* Inhibit a bunch of warnings
Emacs is very chatty by default. That's probably a good thing, but I just don't care about these warnings.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------