Update the default font and the way it is set
Why is this change needed?
--------------------------
- I wanted a different font
  - I wanted that font to be convenient to install and look good to me
- I want it to iteratively look for fonts and choose the most preferential one

How does it address the issue?
------------------------------
I changed the default to Fira Code, which is available via homebrew. I didn't finish writing code that will look for fonts, but I did conditionalize it so it won't complain if you don't have Fira.

The ideal, in order, would be for it to
- automatically install the default font before using it; I'm hoping to write a generic homebrew_install function similar to our usual script/setup process.
- look for a series of fonts, use the best one it could find
- warn you it can't find the default font and tell you what to do about it
- this
- the thing from before, where it failed and complained the whole way

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------