Install doom modeline
Why is this change needed?
--------------------------
The default modeline shows a lot of information I don't think is helpful and isn't visually pleasing.

How does it address the issue?
------------------------------
Doom uses icons for file types and emphasizes the file path and cursor position (the two things I think are the most important). It can be customized to be a lot more useful, but I did not do that.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------