Move all of the global changes under a "Global" header
Why is this change needed?
--------------------------
I want the config to be easy to navigate as it grows.

How does it address the issue?
------------------------------
The current heirarchy is intended to be

- Global - stuff that changes everything
- Per-Mode - stuff that changes one more or a few; place to install a new mode you need
- Personal - things that other people don't like, but you do

This change moves a bunch of stuff under the Global heading. That section needs more organizing, but it's becoming reasonable.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------