Trying out a new git differ
This has fancier output that looks more like github compare views. The
link below has screenshots.

I think this also uses `bat` which looks pretty cool so I've included it
here. In order to fix an issue along the way I needed a version of
`less` more up-to-date than the Mojave default, so `less` is added to
the `Brewfile`.

https://github.com/dandavison/delta