Trying out a new shell file management tool, sheldon
This looks like an interesting and composable way to manage components
of shell files. The main thing is allowing to point to a repo (or
directory) and specify the loading of individual files from that repo.
This allows fine control over what code is loaded and an ability to
reference the source and get updates.

This includes an example plugin using a local directory. It's just a
simple alias to illustrate how this can work. I'll try making some
future changes using sheldon plugins as well to try to get a feel for
sheldon.

Note that when you add local directory plugins you'll need to run the
stow setup before they'll exist where sheldon looks for them.

https://github.com/rossmacarthur/sheldon