Remove Brewfile extries that cause issues
The script/setup files handle mysql and postgres, and use specific
versions as well. If this Brewfile ran first it could cause conflicts.
Most recently the unversioned mysql put a config in place that wouldn't
allow 5.7 to start.

Java is now handled through asdf.

XCode shouldn't be pegged to this specific version, but instead should
use whatever XCode version is in the App Store and is appropriate to the
current OS version.

https://stackoverflow.com/questions/58245118/warning-the-post-install-step-did-not-complete-successfully-when-trying-to-ins