Move `lc` to an executable and add `--clear`
This allows modifications to be picked up without reloading the shell.
The speed difference shouldn't be an issue because it's just run once
per command in general.

This moves it to a sheldon plugin as well, since I think that might be
the future of configuration.

This also adds a `--clear` option to clear notifications that are
showing. This was a handy feature of Growl that I've been missing. I
connected this to Ctrl-Opt-Cmd-C using iCanHazShortcut.