Fix a brew more brew bundle issues
Append `.Brewfile` to filenames so Atom will syntax highlight them.

Comment out geeknote for now since it's broken, though I'd like to bring
it back in the future.

Update osxfuse to macfuse which is the newer version. Use a different
sshfs because the old one was removed from core since it depended on
macfuse which didn't have a compatible license.

Switch the cask tab. The old code threw a warning and told you to switch
to this, so let's do it.

Update the install wrapper so it won't incidentally upgrade packages.
Without this, if there's a newer version of a package it will get
upgraded. In general we only want to upgrade things intentionally.