Add an "About" section to config.org
Why is this change needed?
--------------------------
It's a community config, so there should be some guidelines about
- what users can expect when using the config
- what contributions are likely to be accepted

How does it address the issue?
------------------------------
I titled it "About" because it seemed like an easy place to
start. It's intended to be a starting place, but it's probably not the
end.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/14355728/todos/2209223512