Add documentation to the config
Why is this change needed?
--------------------------
One of the primary goals of the config is to have every configuration
change documented so future-you and anybody else that reads it
understands why that config exists instead of just what it does.

How does it address the issue?
------------------------------
I expanded the description of a few things and split the generic
"clean ui" group into separate entries that were easier to document.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------