Add org roam and org roam ui
Why is this change needed?
--------------------------
I want a better way of recording and organizing knowledge. I reviewed
a bunch of tools in this space including, but not limited to,

- roam
- logseq
- notion
- obsidian

I was surprised to find that Emacs provided the best tooling
overall. I had a lot of requirements, but the key ones were

- no writing style requirements - for example, Roam and Logseq require
  everything to be a node in a tree. I just want to start writing
  without any thought to format. Also, a tree is not always the best
  way to represent information.
- stored in plain text - everything I write should be as portable as
  possible
- stored in a place I can control or export - if I decide to switch
  tools or the tool I'm using goes away, that shouldn't make me start over
- navigate by reference or tag - it should be easy to find linked or
  unlinked references to anything I write about, either because I'm
  implicitly mentioned it somewhere, I've explicitly linked it, or
  I've explicitly grouped the ideas together with a tag

How does it address the issue?
------------------------------
Org Roam is the most popular zettelkasten package for Emacs and does
the above once configured correctly. I also added Org Roam UI, a web
interface for navigating your notes, to make it even easier to see
knowledge relations and find notes I'm looking for.

This implementation doesn't address all of my needs, but committing
this is the last thing I need to do to switch laptops so it's
happening anyway. In particular, I haven't

- made sure that everything that's embedded in a note is stored in the
  note repository (e.g. adding an image should put the image in the
  repo)
- actually turned the org roam directory into a repo for maximum
  storage safety
- installed a flashcard solution
- enabled transclusion
- decided on a publishing feature (if I want one)

Still, it should fire up and work just fine for anyone. I'm already
using it fairly extensively. It might even eventually take the place
of my tmp file.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------