Remove custom org roam method for immediate insertion
Why is this change needed?
--------------------------
It was recommended by several people while I was setting it up, but I
never want it.

How does it address the issue?
------------------------------
I'm deleting it because it's bad to have code you never use.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/14355728/todos/4853961505