Switch Jason's EDITOR to vscode
Why is this change needed?
--------------------------
VSCode is faster and seems to be getting more development than Atom.
I've been using it for awhile and have been enjoying it so I'll make it
offical.

How does it address the issue?
------------------------------
Switch the EDITOR variable to a newly added script to handle opening the
editor and waiting for the file to be closed.

This also adds a comment to help me waste less time next time trying a
simpler strategy. This avoids issues like:
```
$ EDITOR="code --wait" crontab -e
crontab: code --wait: No such file or directory
crontab: "code --wait" exited with status 1
```