Fix path issues with lc
Why is this change needed?
--------------------------
This needs to know about homebrew to find terminal-notifier. The edge
case is when run through iCanHazShortcut, which runs the script outside
of a normal shell, and so doesn't know about homebrew.

How does it address the issue?
------------------------------
This adds homebrew to the path. We need to be a little careful about the
position since we want nodenv to take presedence over homebrew when run
from a normal shell.

Maybe there's a better way to do this, perhaps some way to give this
script a normal shell evnironment like a terminal would have, but this
is easy for the moment and won't have speed implications.