Add Silver Searcher as an option for projectile searching
Why is this change needed?
--------------------------
The second most common thing I do with Projectile after using it to
find a file in a project is search the project for instances of a
thing. I was using grep because it's easy, but I always use `ag` in my
terminal so it makes sense to expand that option to Emacs.

How does it address the issue?
------------------------------
Add the package

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------