Add an alias that pushes a git branch and sets local to track the remote
Why is this change needed?
--------------------------
It's the most common thing in the world. You make a new branch, then
you want to push those changes to the remote and you get that message
about needing to set the tracking branch or specify more info when you
push. Most documentation about this makes it sound like it's very
straightforward to automate, but I kept running into cases where the
simple solutions didn't work and I'd have to type things out the long
way.

How does it address the issue?
------------------------------
Add the alias 'gpu' that does all of the steps idempotently. If you
haven't pushed and set tracking, it does it for you. If that's already
done, it pushes. I left a link to the most relevant SO in case the
best way of doing this changes in the future.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------