Add a method for changing the Org Roam directory
Why is this change needed?
--------------------------
I use Org Roam to manage my content and notes for the DnD campaigns
I'm in. The campaigns often have overlapping locations and NPCs, but
differ in what happened in those locations and with those NPCs. Having
all campaign notes in the same Org Roam directory makes
differentiating them difficult. It also pollutes the non-dnd notes
with lots and lots of unrelated notes that won't ever link to anything
else. Surprisingly, it can be difficult to tell the difference between
the name of a cocktail and the name of a place in DnD.

How does it address the issue?
------------------------------
I tried differentiating with elaborate note-name prefixing and/or
tagging, but it was tedious and error prone. Instead, I've added the
`eb/org-roam-change-directory` method for changing the target note
directory while Emacs is running. It's kind of the bare minimum. To
make it more helpful to me immediately, I added some helper methods in
my personal config to switch between the ones I care about.

Some improvements I'd like to make in the future:
- user can set a variable that the function reads to provide directory
  selection options without needing to make their own functions
- function remembers recently selected directories, so if you do it
  the tedious way one time it will be easy in the future (like
  projectile finding projects you care about)
- function should prompt you for a directory path if you don't provide
  one instead of making you know the syntax the emacs syntax for
  providing input to interactive functions

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------