Change default dailies template so the title is human friendly
Why is this change needed?
--------------------------
I interlink my notes a lot, even the dailies. Example: In my DnD
campaigns I have a daily for the current session but take most of
the notes in the location files. To keep track of when something
happens in real life, I insert the date before I start taking notes in
a particular location. That has the added benefit of making it
possible to go to the daily for that day to see all of the locations
the characters were in by looking at the backlinks. Hooray automatic
inference!

How does it address the issue?
------------------------------
I overrode the default daily capture template.

It was actually kind of hard to figure out what syntax I wanted. The
OrgRoam docs give you an example over overriding the template[1], but
don't give you format ideas. Org documentation tells you what formats
you can use[2], but the list includes ‘%<FORMAT>’ and it doesn't link
to the valid values of that. The valid values of FORMAT are defined in
the Emacs documentation[3].

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - https://www.orgroam.com/manual.html#org_002droam_002ddailies
2 - https://orgmode.org/manual/Template-expansion.html
3 - https://www.gnu.org/software/emacs/manual/html_node/elisp/Time-Parsing.html