Stop automatically indenting org text to the level of the node
Why is this change needed?
--------------------------
As previously configured, text beneath a node in org would
automatically be indented to the level of the node, so it might look
like

*
 foo
**
  foo
***
   foo

I think that's maddening

How does it address the issue?
------------------------------
The org-adapt-indentation controls that feature. I turned it off.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------