Delete install.sh
Why is this change needed?
--------------------------
It's yet another install script! There are so many and it's
confusing. I'm consolidating them all into a single architecture.

How does it address the issue?
------------------------------
The entire functionality of this has been moved to
script/install/dotfiles which is called from script/setup and can also
be called directly.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------