Remove specific reference to code in dropbox setup
Why is this change needed?
--------------------------
The reference to code is misleading. At this point, most people that
use Dropbox have lots of files and are better off starting their sync
as soon as possible regardless of what they have in it.

How does it address the issue?
------------------------------
Change the language to encourage syncing regardless of what's in dropbox.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------