Add a "What" section to the README
Why is this change needed?
--------------------------
It stood out to me when reviewing the README on GitHub. If we want to share this with other people, they're going to need to know what it is.

How does it address the issue?
------------------------------
Explain the common files.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------