Add directions for storing secrets in an encrypted volume
Why is this change needed?
--------------------------
It's important to be able to access your secrets and related
configuration quickly, even in the event of a catastrophic computer
failure. However, it's not safe to store naked secrets on a web service.

How does it address the issue?
------------------------------
Add directions for creating an encrypted volume that's automatically
mounted on boot and contains the secrets and configuration. Then,
symlink those secrets from the volume to the correct places on the
actual filesystem.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------