Add a function to communicate with chatgpt on the command line
Why is this change needed?
--------------------------
It's an experiment in accelerating development with the convenient use
of generative AI. There are currently LOTS of articles about it. This
implementation is based on this article: https://kadekillary.work/posts/1000x-eng/

How does it address the issue?
------------------------------
It adds a bash function, ask_chatgpt, that takes a prompt in the same
way chatgpt would in the browser. I also aliased it to "h". I'm not
sure I think that's the best alias, but it works for now.

I don't actually expect this function to be all that useful by
itself. I think a key feature of chatgpt is that it remembers what
you're talking to it about, so you can make modifications and
adjustments. That doesn't happen using this command line
interface. However, I expect it will be a lot easier to automate
things and provide large bodies of input to chatgpt over command
line. Maybe even setting large bodies of context before asking a question.

Per discussion about making the common files more development
friendly, I added a line to the top of bashrc that sources all the
bash files in a directory. The directory is hardcoded right now in a
way that's probably not going to work for other people, but I had to
stop shaving at some point.

Anyway, the idea is that ALL function extension for bash will be done
in files in that directory[1].

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - https://3.basecamp.com/3093825/buckets/28354346/messages/5105707052#__recording_5163658670