Add some error handling
Why is this change needed?
--------------------------
I got tired of it printing "null" whenever there was a problem with my request.

How does it address the issue?
------------------------------
Check to see if they API returns an error message. Print that message
if there is one.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------