Add a function generating images from OpenAI
Why is this change needed?
--------------------------
I want to put weird things in channel.

How does it address the issue?
------------------------------
There's now a function called img_gpt in the terminal that takes a
text description and provides a url for the created image. I aliased
it to "hi", since the basic chat function is "h". And I think "hi" is funny.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://kadekillary.work/posts/1000x-eng/