Prevent asdf error before java is installed
Why is this change needed?
--------------------------
I believe I was getting an error message during computer setup because
the java plugin wasn't installed yet. This resulted in some weird text
in the PATH (an error message from asdf).

How does it address the issue?
------------------------------
Check if the asdf java plugin is installed before setting the JAVA_HOME.