Add some personal brew stuff
Why is this change needed?
--------------------------
Fswatch is a nice tool for running commands when files change. I've used
it to run tests on save.

Slack is out chat tool, so it could probably go in the shared file, but
I'm just trying to quickly commit outstanding stuff at the moment.

Google has a free VPN service, so I'm trying it out.