Remove Arq from the general Brewfile
Why is this change needed?
--------------------------
Folks complained about Arq hassling them when they weren't using it.

How does it address the issue?
------------------------------
I already have it in my personal brewfiles, so it doesn't need to be in
the general one. I like it, but having it annoy people isn't going to
get them interested.