Update stow to better handle newly appearing files
Why is this change needed?
--------------------------
Since foxtrot died it became obvious that I should have been saving more
of my dotfiles in a place that was easier to recover. It would have been
nice if the common files would have alerted me when new files showed up
that I might want to backup or commit to the common files.

How does it address the issue?
------------------------------
Change the way we stow so more directories are symlinked if they can be,
so new files are more likely to show up in a more obvious place.

Also check to see if there are new files that look like they might be
good candidates to move into the common files repo.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?